Extracting display manager definition
